+++ packages/tabs/src/TabContext.tsx
@@ -1,13 +1,12 @@
 import * as React from 'react';
 
 export interface TabContextState {
-  tabCount: number;
   onClick: (key: number) => void;
   selectedIndex: number;
+  styles?: React.CSSProperties;
 }
 
 const TabContext = React.createContext<TabContextState>({
-  tabCount: 0,
   onClick: () => true,
   selectedIndex: 0,
 });